backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
ecdsa==0.19.1
fastapi==0.128.4
httpx==0.27.0
greenlet==3.3.1
h11==0.16.0
httptools==0.7.1
idna==3.11
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.0.0
python-jose==3.4.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
sqlmodel==0.0.32
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.27.0
watchfiles==1.1.1
websockets==16.0
